+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - **[command_advice]** improvements and fixes
 - **feature:** optimize the default config schema.
 - **fix:** make it working in MC version [1.20, 1.20.2].
+- **feature:** Pre-compile the regex pattern to improve performance.